fix: Use common filters across platform adapters
