+++ consumer/src/lib.rs
@@ -9,8 +9,12 @@ pub use tree::{ChangeHandler as TreeChangeHandler, State as TreeState, Tree};
 pub(crate) mod node;
 pub use node::{DetachedNode, Node, NodeState};
 
+pub(crate) mod filters;
+pub use filters::{
+    common_filter, common_filter_detached, common_filter_with_root_exception, FilterResult,
+};
+
 pub(crate) mod iterators;
-pub use iterators::FilterResult;
 
 pub(crate) mod text;
 pub use text::{
