+++ platforms/macos/src/filters.rs
@@ -1,0 +1,8 @@
+// Copyright 2023 The AccessKit Authors. All rights reserved.
+// Licensed under the Apache License, Version 2.0 (found in
+// the LICENSE-APACHE file) or the MIT license (found in
+// the LICENSE-MIT file), at your option.
+
+pub(crate) use accesskit_consumer::{
+    common_filter as filter, common_filter_detached as filter_detached,
+};
